Move PostsIterator class to a separate file
